requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
python-decouple==3.8
Pillow==10.0.1
django-timezone-field==6.0.1
gunicorn==21.2.0
whitenoise>=6.5.0